+++ libraries/Update/examples/Signed_OTA_Update/Signed_OTA_Update.ino
@@ -92,8 +92,8 @@ void performOTAUpdate() {
   const size_t signatureSize = 512;
   size_t firmwareSize = contentLength - signatureSize;
 
-  Serial.printf("Actual firmware size: %zu bytes\n", firmwareSize);
-  Serial.printf("Signature size: %zu bytes\n", signatureSize);
+  Serial.printf("Actual firmware size: %lu bytes\n", (unsigned long)firmwareSize);
+  Serial.printf("Signature size: %lu bytes\n", (unsigned long)signatureSize);
 
   // Select hash algorithm
 #ifdef USE_SHA256
@@ -171,7 +171,7 @@ void performOTAUpdate() {
     delay(1);
   }
 
-  Serial.printf("Written: %zu bytes\n", written);
+  Serial.printf("Written: %lu bytes\n", (unsigned long)written);
 
   // End the update - this will verify the signature
   if (Update.end()) {
